+++ org.xtext.example.mydsl.interpreter/src/org/xtext/example/mydsl/debugger/processing/expression/ConstantVariableExpressionExecutor.java
@@ -29,7 +29,7 @@ public class ConstantVariableExpressionExecutor extends AbstractStackHelper impl
         Object value = null;
         int size = 0;
         if(expression.getDimension() != null && expression.getDimension().getSize() > 0) {
-            size = expression.getDimension().getSize();
+            size = (int) expression.getDimension().getSize();
             Object array[] = generateArrayValue(type, size);
             value = array;
         } else {
